+++ apollo-composition/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.4.0 - Unreleased
 
+- Update `apollo-federation` dependency to v2.11.0-abstract.1
+
 - Update `apollo-federation` dependency to v2.8.0-preview.0 (from v2.7.0)
 
 - Update `apollo-federation` dependency to v2.7.0 (from v2.5.0)