+++ apollo-federation-types/CHANGELOG.md
@@ -4,6 +4,8 @@ Not every version is listed here because versions before 0.14.0 did not have a c
 
 ## 0.16.0 - Unreleased
 
+- Update `apollo-federation` dependency to v2.11.0-abstract.1
+
 - Update `apollo-federation` dependency to v2.8.0-preview.0 (from v2.7.0)
 
 - Update `apollo-federation` dependency to v2.7.0 (from v2.5.0)